Crate rusoto_translate

Provides translation between one source language and another of the same set of languages.

If you're using the service, you're probably looking for TranslateClient and Translate.

Structs

AppliedTerminology
The custom terminology applied to the input text by Amazon Translate for the translated text response. This is optional in the response and will only be present if you specified terminology input in the request. Currently, only one terminology can be applied per TranslateText request.

DeleteTerminologyRequest

EncryptionKey
The encryption key used to encrypt the custom terminologies used by Amazon Translate.

GetTerminologyRequest

GetTerminologyResponse

ImportTerminologyRequest

ImportTerminologyResponse

ListTerminologiesRequest

ListTerminologiesResponse

Term
The term being translated by the custom terminology.

TerminologyData
The data associated with the custom terminology.

TerminologyDataLocation
The location of the custom terminology data.

TerminologyProperties
The properties of the custom terminology.

TranslateClient
A client for the Amazon Translate API.

TranslateTextRequest

TranslateTextResponse

Enums

DeleteTerminologyError
Errors returned by DeleteTerminology

GetTerminologyError
Errors returned by GetTerminology

ImportTerminologyError
Errors returned by ImportTerminology

ListTerminologiesError
Errors returned by ListTerminologies

TranslateTextError
Errors returned by TranslateText

Traits

Translate
Trait representing the capabilities of the Amazon Translate API. Amazon Translate clients implement this trait.