Trait rusoto_translate::Translate

pub trait Translate {
    fn delete_terminology(
        &self, 
        input: DeleteTerminologyRequest
    ) -> RusotoFuture<(), DeleteTerminologyError>;

    fn get_terminology(
        &self, 
        input: GetTerminologyRequest
    ) -> RusotoFuture<GetTerminologyResponse, GetTerminologyError>;

    fn import_terminology(
        &self, 
        input: ImportTerminologyRequest
    ) -> RusotoFuture<ImportTerminologyResponse, ImportTerminologyError>;

    fn list_terminologies(
        &self, 
        input: ListTerminologiesRequest
    ) -> RusotoFuture<ListTerminologiesResponse, ListTerminologiesError>;

    fn translate_text(
        &self, 
        input: TranslateTextRequest
    ) -> RusotoFuture<TranslateTextResponse, TranslateTextError>;
}

Trait representing the capabilities of the Amazon Translate API. Amazon Translate clients implement this trait.

Required methods

fn delete_terminology(
&self,
input: DeleteTerminologyRequest
) -> RusotoFuture<(), DeleteTerminologyError>

A synchronous action that deletes a custom terminology.

fn get_terminology(
&self,
input: GetTerminologyRequest
) -> RusotoFuture<GetTerminologyResponse, GetTerminologyError>

Retrieves a custom terminology.

fn import_terminology(
&self,
input: ImportTerminologyRequest
) -> RusotoFuture<ImportTerminologyResponse, ImportTerminologyError>

Creates or updates a custom terminology, depending on whether or not one already exists for the given terminology name. Importing a terminology with the same name as an existing one will merge the terminologies based on the chosen merge strategy. Currently, the only supported merge strategy is OVERWRITE, and so the imported terminology will overwrite an existing terminology of the same name.

If you import a terminology that overwrites an existing one, the new terminology take up to 10 minutes to fully propagate and be available for use in a translation due to cache policies with the DataPlane service that performs the translations.

fn list_terminologies(
&self,
input: ListTerminologiesRequest
) -> RusotoFuture<ListTerminologiesResponse, ListTerminologiesError>

Provides a list of custom terminologies associated with your account.

fn translate_text(
&self,
input: TranslateTextRequest
) -> RusotoFuture<TranslateTextResponse, TranslateTextError>

Translates input text from the source language to the target language. It is not necessary to use English (en) as either the source or the target language but not all language combinations are supported by Amazon Translate. For more information, see Supported Language Pairs.

Arabic (ar)

Chinese (Simplified) (zh)

Chinese (Traditional) (zh-TW)

Czech (cs)

Danish (da)

Dutch (nl)

English (en)

Finnish (fi)

French (fr)

German (de)

Hebrew (he)

Indonesian (id)

Italian (it)

Japanese (ja)

Korean (ko)

Polish (pl)

Portuguese (pt)

Russian (ru)

Spanish (es)

Swedish (sv)

Turkish (tr)

To have Amazon Translate determine the source language of your text, you can specify auto in the SourceLanguageCode field. If you specify auto, Amazon Translate will call Amazon Comprehend to determine the source language.

Loading content...